Refactors ComponentManager as initComponents
After removing the manifest, keeping ComponentManager as a class felt forced. IMHO it works better as a simple suite of functions.
